CS50 Scratch Project
Scratch project 755922271 (https://scratch.mit.edu/projects/755922271/) is designed to be an entertaining supplement to David's introduction to algorithms where he tears a phonebook in half to illustrate how an efficient algorithm can be designed. In the Scratch version, the David Sprite needs to 'catch and hold' (i.e., touch for at least 2 seconds) the Phonebook sprite in order to open the Phonebook. Once the Phonebook has been secured and opened, the user must press the space key for a fixed period to 'tear' the book and in turn complete the Phonebook Algorithm example.